refactor: clean up and refine code

+++ src/controllers/member-controller.ts
@@ -1,10 +1,10 @@
 import { Request, Response, NextFunction } from 'express'
 import fs from 'fs'
 
-import { Member } from '../models'
-import { addMemberSchema, editMemberSchema } from '../schemas'
-import { ErrorType } from '../types'
-import { getApiUrl, getDefaultImagePath, getImagePath } from '../util'
+import { Member } from 'models'
+import { addMemberSchema, editMemberSchema } from 'schemas'
+import { ErrorType } from 'types'
+import { getApiUrl, getDefaultImagePath, getImagePath } from 'util'
 
 export const getMembers = async (_: Request, res: Response, next: NextFunction) => {
   try {
@@ -18,7 +18,13 @@ export const getMembers = async (_: Request, res: Response, next: NextFunction) 
 
 export const getMember = async (req: Request, res: Response, next: NextFunction) => {
   try {
-    const member = await Member.findById(req.body.id).select('-__v')
+    if(!req.params.id.match(/^[a-f\d]{24}$/i)) {
+      const error: ErrorType = new Error("Invalid id.")
+      error.statusCode = 422
+      throw error
+    }
+
+    const member = await Member.findById(req.params.id).select('-__v')
 
     if(!member) {
       const error: ErrorType = new Error("Member with this id not found.")
@@ -53,6 +59,12 @@ export const addMember = async (req: Request, res: Response, next: NextFunction)
 
 export const changeMemberAvatar = async (req: Request, res: Response, next: NextFunction) => {
   try {
+    if(!req.body.memberId.match(/^[a-f\d]{24}$/i)) {
+      const error: ErrorType = new Error("Invalid id.")
+      error.statusCode = 422
+      throw error
+    }
+    
     const member = await Member.findById(req.body.memberId)
 
     if(!req.file) {
@@ -92,6 +104,12 @@ export const changeMemberAvatar = async (req: Request, res: Response, next: Next
 export const editMember = async (req: Request, res: Response, next: NextFunction) => {
   try {
     await editMemberSchema.validateAsync(req.body)
+    
+    if(!req.body.id.match(/^[a-f\d]{24}$/i)) {
+      const error: ErrorType = new Error("Invalid id.")
+      error.statusCode = 422
+      throw error
+    }
 
     const member = await Member.findOneAndUpdate(
       {
@@ -106,7 +124,7 @@ export const editMember = async (req: Request, res: Response, next: NextFunction
       throw error
     }
 
-    res.status(201).json({
+    res.status(200).json({
       message: 'Member updated successfully!',
     })
   } catch (err) {
@@ -116,6 +134,12 @@ export const editMember = async (req: Request, res: Response, next: NextFunction
 
 export const deleteMember = async (req: Request, res: Response, next: NextFunction) => {
   try {
+    if(!req.body.id.match(/^[a-f\d]{24}$/i)) {
+      const error: ErrorType = new Error("Invalid id.")
+      error.statusCode = 422
+      throw error
+    }
+
     const member = await Member.findByIdAndRemove(req.body.id)
 
     if (!member) {
